composite-video-parser/half-digit-classifier/train/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
onnx>=1.16.0
onnxruntime>=1.17.0
Pillow>=10.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
